allow api/share public access
