+++ src/app/s/[slug]/page.js
